+++ plugins/clutch-wp/clutch.php
@@ -7,7 +7,7 @@
  * License:           GPLv2 or later
  * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
  *
- * Version:           2.4.2
+ * Version:           2.4.11
  * Requires PHP:      7.4
  * Requires at least: 5.7
  *