+++ plugins/clutch-wp/includes/blocks/assets/clutch-gen/index.asset.php
@@ -1,12 +1,1 @@
-<?php return [
-	'dependencies' => [
-		'react-jsx-runtime',
-		'wp-block-editor',
-		'wp-blocks',
-		'wp-components',
-		'wp-data',
-		'wp-i18n',
-		'wp-primitives',
-	],
-	'version' => 'f2952c8bc99dd43c3f71',
-];
+<?php return array('dependencies' => array('react-jsx-runtime', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-data', 'wp-element', 'wp-i18n', 'wp-primitives'), 'version' => 'a7742b942dc980ee649d');