+++ packages/react/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @clutch-wp/react
 
+## 1.1.12
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.12
+
+## 1.1.11
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.11
+
+## 1.1.10
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.10
+
+## 1.1.9
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.9
+
+## 1.1.8
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.8
+
+## 1.1.7
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.7
+
+## 1.1.6
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.6
+
+## 1.1.5
+
+### Patch Changes
+
+- @clutch-wp/sdk@1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [9af549e]
+  - @clutch-wp/sdk@1.1.4
+
 ## 1.1.3
 
 ### Patch Changes