+++ plugins/clutch-wp/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @clutch-wp/clutch-plugin
 
+## 2.4.11
+
+### Patch Changes
+
+- 304291a: Fix incorrect sanitization of component ids
+
+## 2.4.10
+
+### Patch Changes
+
+- 64242bd: Fix "children" named slots in custom blocks not being processed
+
+## 2.4.9
+
+### Patch Changes
+
+- cccc6ed: Recompute websites array keys on deletion
+
+## 2.4.8
+
+### Patch Changes
+
+- a79c954: List connected local projects
+
+## 2.4.7
+
+### Patch Changes
+
+- 44c20bf: Fix ClutchAdminBar error on missing deploymentId
+
+## 2.4.6
+
+### Patch Changes
+
+- 740c078: Fix ClutchAdminBar websites dropdown
+
+## 2.4.5
+
+### Patch Changes
+
+- 40793c6: Add slot support for custom blocks
+
+## 2.4.4
+
+### Patch Changes
+
+- 0d4dbbc: Improve performance through ISR request batching
+
+## 2.4.3
+
+### Patch Changes
+
+- 9af549e: Add variants and string array property support for custom blocks
+
 ## 2.4.2
 
 ### Patch Changes